fix(docs): fix salesforce docs & update styling
