+++ apps/docs/app/[lang]/[[...slug]]/page.tsx
@@ -182,7 +182,11 @@ export default async function Page(props: { params: Promise<{ slug?: string[]; l
         tableOfContent={{
           style: 'clerk',
           enabled: true,
-          header: <div className='mb-2 font-medium text-sm'>On this page</div>,
+          header: (
+            <div key='toc-header' className='mb-2 font-medium text-sm'>
+              On this page
+            </div>
+          ),
           footer: <TOCFooter />,
           single: false,
         }}
