+++ apps/docs/components/ui/icon-mapping.ts
@@ -116,114 +116,114 @@ import {
 type IconComponent = ComponentType<SVGProps<SVGSVGElement>>
 
 export const blockTypeToIconMap: Record<string, IconComponent> = {
-  calendly: CalendlyIcon,
-  mailchimp: MailchimpIcon,
-  postgresql: PostgresIcon,
-  twilio_voice: TwilioIcon,
-  elasticsearch: ElasticsearchIcon,
-  rds: RDSIcon,
-  translate: TranslateIcon,
-  dynamodb: DynamoDBIcon,
-  wordpress: WordpressIcon,
-  tavily: TavilyIcon,
-  zendesk: ZendeskIcon,
-  stagehand_agent: StagehandIcon,
-  youtube: YouTubeIcon,
-  supabase: SupabaseIcon,
-  vision: EyeIcon,
   zoom: ZoomIcon,
-  confluence: ConfluenceIcon,
-  arxiv: ArxivIcon,
-  webflow: WebflowIcon,
-  pinecone: PineconeIcon,
-  apollo: ApolloIcon,
-  whatsapp: WhatsAppIcon,
-  typeform: TypeformIcon,
-  qdrant: QdrantIcon,
-  shopify: ShopifyIcon,
-  asana: AsanaIcon,
-  apify: ApifyIcon,
-  memory: BrainIcon,
-  gitlab: GitLabIcon,
-  polymarket: PolymarketIcon,
-  serper: SerperIcon,
-  linear: LinearIcon,
-  exa: ExaAIIcon,
-  telegram: TelegramIcon,
-  salesforce: SalesforceIcon,
-  hubspot: HubspotIcon,
-  hunter: HunterIOIcon,
-  linkup: LinkupIcon,
-  mongodb: MongoDBIcon,
-  pylon: PylonIcon,
-  airtable: AirtableIcon,
-  discord: DiscordIcon,
-  ahrefs: AhrefsIcon,
-  neo4j: Neo4jIcon,
-  tts: TTSIcon,
-  jina: JinaAIIcon,
-  google_docs: GoogleDocsIcon,
-  perplexity: PerplexityIcon,
-  google_search: GoogleIcon,
-  x: xIcon,
-  kalshi: KalshiIcon,
-  google_calendar: GoogleCalendarIcon,
   zep: ZepIcon,
-  posthog: PosthogIcon,
-  grafana: GrafanaIcon,
-  google_slides: GoogleSlidesIcon,
-  microsoft_planner: MicrosoftPlannerIcon,
-  thinking: BrainIcon,
-  pipedrive: PipedriveIcon,
-  dropbox: DropboxIcon,
-  stagehand: StagehandIcon,
-  google_forms: GoogleFormsIcon,
-  file: DocumentIcon,
-  mistral_parse: MistralIcon,
-  gmail: GmailIcon,
-  openai: OpenAIIcon,
-  outlook: OutlookIcon,
-  incidentio: IncidentioIcon,
-  onedrive: MicrosoftOneDriveIcon,
-  resend: ResendIcon,
-  google_vault: GoogleVaultIcon,
-  sharepoint: MicrosoftSharepointIcon,
-  huggingface: HuggingFaceIcon,
-  sendgrid: SendgridIcon,
-  video_generator: VideoIcon,
-  smtp: SmtpIcon,
-  google_groups: GoogleGroupsIcon,
-  mailgun: MailgunIcon,
-  clay: ClayIcon,
-  jira: JiraIcon,
-  search: SearchIcon,
-  linkedin: LinkedInIcon,
-  wealthbox: WealthboxIcon,
-  notion: NotionIcon,
-  elevenlabs: ElevenLabsIcon,
-  microsoft_teams: MicrosoftTeamsIcon,
-  github: GithubIcon,
-  ssh: SshIcon,
-  google_drive: GoogleDriveIcon,
-  sentry: SentryIcon,
-  reddit: RedditIcon,
-  parallel_ai: ParallelIcon,
-  stripe: StripeIcon,
-  s3: S3Icon,
-  trello: TrelloIcon,
-  mem0: Mem0Icon,
-  knowledge: PackageSearchIcon,
-  intercom: IntercomIcon,
-  twilio_sms: TwilioIcon,
-  slack: SlackIcon,
-  datadog: DatadogIcon,
-  microsoft_excel: MicrosoftExcelIcon,
-  image_generator: ImageIcon,
-  google_sheets: GoogleSheetsIcon,
+  zendesk: ZendeskIcon,
+  youtube: YouTubeIcon,
+  x: xIcon,
+  wordpress: WordpressIcon,
   wikipedia: WikipediaIcon,
-  cursor: CursorIcon,
-  firecrawl: FirecrawlIcon,
-  mysql: MySQLIcon,
-  browser_use: BrowserUseIcon,
+  whatsapp: WhatsAppIcon,
+  webflow: WebflowIcon,
+  wealthbox: WealthboxIcon,
+  vision: EyeIcon,
+  video_generator: VideoIcon,
+  typeform: TypeformIcon,
+  twilio_voice: TwilioIcon,
+  twilio_sms: TwilioIcon,
+  tts: TTSIcon,
+  trello: TrelloIcon,
+  translate: TranslateIcon,
+  thinking: BrainIcon,
+  telegram: TelegramIcon,
+  tavily: TavilyIcon,
+  supabase: SupabaseIcon,
   stt: STTIcon,
+  stripe: StripeIcon,
+  stagehand_agent: StagehandIcon,
+  stagehand: StagehandIcon,
+  ssh: SshIcon,
+  smtp: SmtpIcon,
+  slack: SlackIcon,
+  shopify: ShopifyIcon,
+  sharepoint: MicrosoftSharepointIcon,
+  serper: SerperIcon,
+  sentry: SentryIcon,
+  sendgrid: SendgridIcon,
+  search: SearchIcon,
+  salesforce: SalesforceIcon,
+  s3: S3Icon,
+  resend: ResendIcon,
+  reddit: RedditIcon,
+  rds: RDSIcon,
+  qdrant: QdrantIcon,
+  pylon: PylonIcon,
+  posthog: PosthogIcon,
+  postgresql: PostgresIcon,
+  polymarket: PolymarketIcon,
+  pipedrive: PipedriveIcon,
+  pinecone: PineconeIcon,
+  perplexity: PerplexityIcon,
+  parallel_ai: ParallelIcon,
+  outlook: OutlookIcon,
+  openai: OpenAIIcon,
+  onedrive: MicrosoftOneDriveIcon,
+  notion: NotionIcon,
+  neo4j: Neo4jIcon,
+  mysql: MySQLIcon,
+  mongodb: MongoDBIcon,
+  mistral_parse: MistralIcon,
+  microsoft_teams: MicrosoftTeamsIcon,
+  microsoft_planner: MicrosoftPlannerIcon,
+  microsoft_excel: MicrosoftExcelIcon,
+  memory: BrainIcon,
+  mem0: Mem0Icon,
+  mailgun: MailgunIcon,
+  mailchimp: MailchimpIcon,
+  linkup: LinkupIcon,
+  linkedin: LinkedInIcon,
+  linear: LinearIcon,
+  knowledge: PackageSearchIcon,
+  kalshi: KalshiIcon,
+  jira: JiraIcon,
+  jina: JinaAIIcon,
+  intercom: IntercomIcon,
+  incidentio: IncidentioIcon,
+  image_generator: ImageIcon,
+  hunter: HunterIOIcon,
+  huggingface: HuggingFaceIcon,
+  hubspot: HubspotIcon,
+  grafana: GrafanaIcon,
+  google_vault: GoogleVaultIcon,
+  google_slides: GoogleSlidesIcon,
+  google_sheets: GoogleSheetsIcon,
+  google_groups: GoogleGroupsIcon,
+  google_forms: GoogleFormsIcon,
+  google_drive: GoogleDriveIcon,
+  google_docs: GoogleDocsIcon,
+  google_calendar: GoogleCalendarIcon,
+  google_search: GoogleIcon,
+  gmail: GmailIcon,
+  gitlab: GitLabIcon,
+  github: GithubIcon,
+  firecrawl: FirecrawlIcon,
+  file: DocumentIcon,
+  exa: ExaAIIcon,
+  elevenlabs: ElevenLabsIcon,
+  elasticsearch: ElasticsearchIcon,
+  dynamodb: DynamoDBIcon,
+  dropbox: DropboxIcon,
+  discord: DiscordIcon,
+  datadog: DatadogIcon,
+  cursor: CursorIcon,
+  confluence: ConfluenceIcon,
+  clay: ClayIcon,
+  calendly: CalendlyIcon,
+  browser_use: BrowserUseIcon,
+  asana: AsanaIcon,
+  arxiv: ArxivIcon,
+  apollo: ApolloIcon,
+  apify: ApifyIcon,
+  airtable: AirtableIcon,
+  ahrefs: AhrefsIcon,
 }
