+++ apps/docs/content/docs/en/introduction/index.mdx
@@ -72,7 +72,7 @@ For custom integrations, leverage our [MCP (Model Context Protocol) support](/mc
   <Video src="introduction/integrations-sidebar.mp4" width={700} height={450} />
 </div>
 
-## AI-Powered Copilot
+## Copilot
 
 **Ask Questions & Get Guidance**  
 Copilot answers questions about Sim, explains your workflows, and provides suggestions for improvements. Use the `@` symbol to reference workflows, blocks, documentation, knowledge, and logs for contextual assistance.
