+++ apps/docs/content/docs/en/knowledgebase/index.mdx
@@ -1,5 +1,6 @@
 ---
-title: Knowledgebase
+title: Overview
+description: Upload, process, and search through your documents with intelligent vector search and chunking
 ---
 
 import { Video } from '@/components/ui/video'
