+++ apps/docs/content/docs/en/knowledgebase/meta.json
@@ -1,0 +1,4 @@
+{
+  "title": "Knowledgebase",
+  "pages": ["index", "tags"]
+}
