+++ apps/sim/tools/salesforce/create_task.ts
@@ -1,0 +1,141 @@
+import type { ToolConfig } from '@/tools/types'
+import { getInstanceUrl } from './utils'
+
+export interface SalesforceCreateTaskParams {
+  accessToken: string
+  idToken?: string
+  instanceUrl?: string
+  subject: string
+  status?: string
+  priority?: string
+  activityDate?: string
+  whoId?: string
+  whatId?: string
+  description?: string
+}
+
+export interface SalesforceCreateTaskResponse {
+  success: boolean
+  output: {
+    id: string
+    success: boolean
+    created: boolean
+    metadata: {
+      operation: 'create_task'
+    }
+  }
+}
+
+export const salesforceCreateTaskTool: ToolConfig<
+  SalesforceCreateTaskParams,
+  SalesforceCreateTaskResponse
+> = {
+  id: 'salesforce_create_task',
+  name: 'Create Task in Salesforce',
+  description: 'Create a new task',
+  version: '1.0.0',
+
+  oauth: {
+    required: true,
+    provider: 'salesforce',
+  },
+
+  params: {
+    accessToken: {
+      type: 'string',
+      required: true,
+      visibility: 'hidden',
+    },
+    idToken: {
+      type: 'string',
+      required: false,
+      visibility: 'hidden',
+    },
+    instanceUrl: {
+      type: 'string',
+      required: false,
+      visibility: 'hidden',
+    },
+    subject: {
+      type: 'string',
+      required: true,
+      visibility: 'user-only',
+      description: 'Task subject (required)',
+    },
+    status: {
+      type: 'string',
+      required: false,
+      visibility: 'user-only',
+      description: 'Status (e.g., Not Started, In Progress, Completed)',
+    },
+    priority: {
+      type: 'string',
+      required: false,
+      visibility: 'user-only',
+      description: 'Priority (e.g., Low, Normal, High)',
+    },
+    activityDate: {
+      type: 'string',
+      required: false,
+      visibility: 'user-only',
+      description: 'Due date YYYY-MM-DD',
+    },
+    whoId: {
+      type: 'string',
+      required: false,
+      visibility: 'user-only',
+      description: 'Related Contact/Lead ID',
+    },
+    whatId: {
+      type: 'string',
+      required: false,
+      visibility: 'user-only',
+      description: 'Related Account/Opportunity ID',
+    },
+    description: {
+      type: 'string',
+      required: false,
+      visibility: 'user-only',
+      description: 'Description',
+    },
+  },
+
+  request: {
+    url: (params) =>
+      `${getInstanceUrl(params.idToken, params.instanceUrl)}/services/data/v59.0/sobjects/Task`,
+    method: 'POST',
+    headers: (params) => ({
+      Authorization: `Bearer ${params.accessToken}`,
+      'Content-Type': 'application/json',
+    }),
+    body: (params) => {
+      const body: Record<string, any> = { Subject: params.subject }
+      if (params.status) body.Status = params.status
+      if (params.priority) body.Priority = params.priority
+      if (params.activityDate) body.ActivityDate = params.activityDate
+      if (params.whoId) body.WhoId = params.whoId
+      if (params.whatId) body.WhatId = params.whatId
+      if (params.description) body.Description = params.description
+      return body
+    },
+  },
+
+  transformResponse: async (response) => {
+    const data = await response.json()
+    if (!response.ok) throw new Error(data[0]?.message || data.message || 'Failed to create task')
+    return {
+      success: true,
+      output: {
+        id: data.id,
+        success: data.success,
+        created: true,
+        metadata: { operation: 'create_task' },
+      },
+    }
+  },
+
+  outputs: {
+    success: { type: 'boolean', description: 'Success' },
+    output: { type: 'object', description: 'Created task' },
+  },
+}
