+++ apps/sim/tools/salesforce/get_dashboard.ts
@@ -1,0 +1,118 @@
+import { createLogger } from '@/lib/logs/console/logger'
+import type { ToolConfig } from '@/tools/types'
+import { extractErrorMessage, getInstanceUrl } from './utils'
+
+const logger = createLogger('SalesforceDashboards')
+
+export interface SalesforceGetDashboardParams {
+  accessToken: string
+  idToken?: string
+  instanceUrl?: string
+  dashboardId: string
+}
+
+export interface SalesforceGetDashboardResponse {
+  success: boolean
+  output: {
+    dashboard: any
+    dashboardId: string
+    components: any[]
+    metadata: {
+      operation: 'get_dashboard'
+      dashboardName?: string
+      folderId?: string
+      runningUser?: any
+    }
+    success: boolean
+  }
+}
+
+/**
+ * Get details for a specific dashboard
+ * @see https://developer.salesforce.com/docs/atlas.en-us.api_analytics.meta/api_analytics/sforce_analytics_rest_api_dashboard_results.htm
+ */
+export const salesforceGetDashboardTool: ToolConfig<
+  SalesforceGetDashboardParams,
+  SalesforceGetDashboardResponse
+> = {
+  id: 'salesforce_get_dashboard',
+  name: 'Get Dashboard from Salesforce',
+  description: 'Get details and results for a specific dashboard',
+  version: '1.0.0',
+
+  oauth: {
+    required: true,
+    provider: 'salesforce',
+  },
+
+  params: {
+    accessToken: { type: 'string', required: true, visibility: 'hidden' },
+    idToken: { type: 'string', required: false, visibility: 'hidden' },
+    instanceUrl: { type: 'string', required: false, visibility: 'hidden' },
+    dashboardId: {
+      type: 'string',
+      required: true,
+      visibility: 'user-only',
+      description: 'Dashboard ID (required)',
+    },
+  },
+
+  request: {
+    url: (params) => {
+      if (!params.dashboardId || params.dashboardId.trim() === '') {
+        throw new Error('Dashboard ID is required. Please provide a valid Salesforce Dashboard ID.')
+      }
+      const instanceUrl = getInstanceUrl(params.idToken, params.instanceUrl)
+      return `${instanceUrl}/services/data/v59.0/analytics/dashboards/${params.dashboardId}`
+    },
+    method: 'GET',
+    headers: (params) => ({
+      Authorization: `Bearer ${params.accessToken}`,
+      'Content-Type': 'application/json',
+    }),
+  },
+
+  transformResponse: async (response, params?) => {
+    const data = await response.json()
+    if (!response.ok) {
+      const errorMessage = extractErrorMessage(
+        data,
+        response.status,
+        `Failed to get dashboard ID: ${params?.dashboardId}`
+      )
+      logger.error('Failed to get dashboard', { data, status: response.status })
+      throw new Error(errorMessage)
+    }
+
+    return {
+      success: true,
+      output: {
+        dashboard: data,
+        dashboardId: params?.dashboardId || '',
+        components: data.componentData || [],
+        metadata: {
+          operation: 'get_dashboard',
+          dashboardName: data.name,
+          folderId: data.folderId,
+          runningUser: data.runningUser,
+        },
+        success: true,
+      },
+    }
+  },
+
+  outputs: {
+    success: { type: 'boolean', description: 'Success status' },
+    output: {
+      type: 'object',
+      description: 'Dashboard data',
+      properties: {
+        dashboard: { type: 'object', description: 'Dashboard details' },
+        dashboardId: { type: 'string', description: 'Dashboard ID' },
+        components: { type: 'array', description: 'Dashboard component data' },
+        metadata: { type: 'object', description: 'Operation metadata' },
+        success: { type: 'boolean', description: 'Operation success status' },
+      },
+    },
+  },
+}
