+++ apps/sim/tools/salesforce/get_leads.ts
@@ -1,0 +1,132 @@
+import type { ToolConfig } from '@/tools/types'
+import { getInstanceUrl } from './utils'
+
+export interface SalesforceGetLeadsParams {
+  accessToken: string
+  idToken?: string
+  instanceUrl?: string
+  leadId?: string
+  limit?: string
+  fields?: string
+  orderBy?: string
+}
+
+export interface SalesforceGetLeadsResponse {
+  success: boolean
+  output: {
+    lead?: any
+    leads?: any[]
+    paging?: {
+      nextRecordsUrl?: string
+      totalSize: number
+      done: boolean
+    }
+    metadata: {
+      operation: 'get_leads'
+      totalReturned?: number
+      hasMore?: boolean
+      singleLead?: boolean
+    }
+    success: boolean
+  }
+}
+
+export const salesforceGetLeadsTool: ToolConfig<
+  SalesforceGetLeadsParams,
+  SalesforceGetLeadsResponse
+> = {
+  id: 'salesforce_get_leads',
+  name: 'Get Leads from Salesforce',
+  description: 'Get lead(s) from Salesforce',
+  version: '1.0.0',
+
+  oauth: {
+    required: true,
+    provider: 'salesforce',
+  },
+
+  params: {
+    accessToken: { type: 'string', required: true, visibility: 'hidden' },
+    idToken: { type: 'string', required: false, visibility: 'hidden' },
+    instanceUrl: { type: 'string', required: false, visibility: 'hidden' },
+    leadId: {
+      type: 'string',
+      required: false,
+      visibility: 'user-only',
+      description: 'Lead ID (optional)',
+    },
+    limit: {
+      type: 'string',
+      required: false,
+      visibility: 'user-only',
+      description: 'Max results (default: 100)',
+    },
+    fields: {
+      type: 'string',
+      required: false,
+      visibility: 'user-only',
+      description: 'Comma-separated fields',
+    },
+    orderBy: {
+      type: 'string',
+      required: false,
+      visibility: 'user-only',
+      description: 'Order by field',
+    },
+  },
+
+  request: {
+    url: (params) => {
+      const instanceUrl = getInstanceUrl(params.idToken, params.instanceUrl)
+      if (params.leadId) {
+        const fields =
+          params.fields || 'Id,FirstName,LastName,Company,Email,Phone,Status,LeadSource'
+        return `${instanceUrl}/services/data/v59.0/sobjects/Lead/${params.leadId}?fields=${fields}`
+      }
+      const limit = params.limit ? Number.parseInt(params.limit) : 100
+      const fields = params.fields || 'Id,FirstName,LastName,Company,Email,Phone,Status,LeadSource'
+      const orderBy = params.orderBy || 'LastName ASC'
+      const query = `SELECT ${fields} FROM Lead ORDER BY ${orderBy} LIMIT ${limit}`
+      return `${instanceUrl}/services/data/v59.0/query?q=${encodeURIComponent(query)}`
+    },
+    method: 'GET',
+    headers: (params) => ({
+      Authorization: `Bearer ${params.accessToken}`,
+      'Content-Type': 'application/json',
+    }),
+  },
+
+  transformResponse: async (response, params?) => {
+    const data = await response.json()
+    if (!response.ok) throw new Error(data[0]?.message || data.message || 'Failed to fetch leads')
+    if (params?.leadId) {
+      return {
+        success: true,
+        output: {
+          lead: data,
+          metadata: { operation: 'get_leads', singleLead: true },
+          success: true,
+        },
+      }
+    }
+    const leads = data.records || []
+    return {
+      success: true,
+      output: {
+        leads,
+        paging: {
+          nextRecordsUrl: data.nextRecordsUrl,
+          totalSize: data.totalSize || leads.length,
+          done: data.done !== false,
+        },
+        metadata: { operation: 'get_leads', totalReturned: leads.length, hasMore: !data.done },
+        success: true,
+      },
+    }
+  },
+
+  outputs: {
+    success: { type: 'boolean', description: 'Success status' },
+    output: { type: 'object', description: 'Lead data' },
+  },
+}
