+++ apps/sim/tools/salesforce/index.ts
@@ -1,51 +1,35 @@
-export {
-  salesforceCreateCaseTool,
-  salesforceDeleteCaseTool,
-  salesforceGetCasesTool,
-  salesforceUpdateCaseTool,
-} from './cases'
-export {
-  salesforceCreateContactTool,
-  salesforceDeleteContactTool,
-  salesforceGetContactsTool,
-  salesforceUpdateContactTool,
-} from './contacts'
 export { salesforceCreateAccountTool } from './create_account'
-export {
-  salesforceGetDashboardTool,
-  salesforceListDashboardsTool,
-  salesforceRefreshDashboardTool,
-} from './dashboards'
+export { salesforceCreateCaseTool } from './create_case'
+export { salesforceCreateContactTool } from './create_contact'
+export { salesforceCreateLeadTool } from './create_lead'
+export { salesforceCreateOpportunityTool } from './create_opportunity'
+export { salesforceCreateTaskTool } from './create_task'
 export { salesforceDeleteAccountTool } from './delete_account'
+export { salesforceDeleteCaseTool } from './delete_case'
+export { salesforceDeleteContactTool } from './delete_contact'
+export { salesforceDeleteLeadTool } from './delete_lead'
+export { salesforceDeleteOpportunityTool } from './delete_opportunity'
+export { salesforceDeleteTaskTool } from './delete_task'
+export { salesforceDescribeObjectTool } from './describe_object'
 export { salesforceGetAccountsTool } from './get_accounts'
-export {
-  salesforceCreateLeadTool,
-  salesforceDeleteLeadTool,
-  salesforceGetLeadsTool,
-  salesforceUpdateLeadTool,
-} from './leads'
-export {
-  salesforceCreateOpportunityTool,
-  salesforceDeleteOpportunityTool,
-  salesforceGetOpportunitiesTool,
-  salesforceUpdateOpportunityTool,
-} from './opportunities'
-export {
-  salesforceDescribeObjectTool,
-  salesforceListObjectsTool,
-  salesforceQueryMoreTool,
-  salesforceQueryTool,
-} from './query'
-export {
-  salesforceGetReportTool,
-  salesforceListReportsTool,
-  salesforceListReportTypesTool,
-  salesforceRunReportTool,
-} from './reports'
-export {
-  salesforceCreateTaskTool,
-  salesforceDeleteTaskTool,
-  salesforceGetTasksTool,
-  salesforceUpdateTaskTool,
-} from './tasks'
+export { salesforceGetCasesTool } from './get_cases'
+export { salesforceGetContactsTool } from './get_contacts'
+export { salesforceGetDashboardTool } from './get_dashboard'
+export { salesforceGetLeadsTool } from './get_leads'
+export { salesforceGetOpportunitiesTool } from './get_opportunities'
+export { salesforceGetReportTool } from './get_report'
+export { salesforceGetTasksTool } from './get_tasks'
+export { salesforceListDashboardsTool } from './list_dashboards'
+export { salesforceListObjectsTool } from './list_objects'
+export { salesforceListReportTypesTool } from './list_report_types'
+export { salesforceListReportsTool } from './list_reports'
+export { salesforceQueryTool } from './query'
+export { salesforceQueryMoreTool } from './query_more'
+export { salesforceRefreshDashboardTool } from './refresh_dashboard'
+export { salesforceRunReportTool } from './run_report'
 export { salesforceUpdateAccountTool } from './update_account'
+export { salesforceUpdateCaseTool } from './update_case'
+export { salesforceUpdateContactTool } from './update_contact'
+export { salesforceUpdateLeadTool } from './update_lead'
+export { salesforceUpdateOpportunityTool } from './update_opportunity'
+export { salesforceUpdateTaskTool } from './update_task'
