+++ apps/sim/tools/salesforce/query_more.ts
@@ -4,23 +4,22 @@ import { extractErrorMessage, getInstanceUrl } from './utils'
 
 const logger = createLogger('SalesforceQuery')
 
-export interface SalesforceQueryParams {
+export interface SalesforceQueryMoreParams {
   accessToken: string
   idToken?: string
   instanceUrl?: string
-  query: string
+  nextRecordsUrl: string
 }
 
-export interface SalesforceQueryResponse {
+export interface SalesforceQueryMoreResponse {
   success: boolean
   output: {
     records: any[]
     totalSize: number
     done: boolean
     nextRecordsUrl?: string
-    query: string
     metadata: {
-      operation: 'query'
+      operation: 'query_more'
       totalReturned: number
       hasMore: boolean
     }
@@ -29,13 +28,16 @@ export interface SalesforceQueryResponse {
 }
 
 /**
- * Execute a custom SOQL query
+ * Retrieve additional query results using the nextRecordsUrl
  * @see https://developer.salesforce.com/docs/atlas.en-us.api_rest.meta/api_rest/dome_query.htm
  */
-export const salesforceQueryTool: ToolConfig<SalesforceQueryParams, SalesforceQueryResponse> = {
-  id: 'salesforce_query',
-  name: 'Run SOQL Query in Salesforce',
-  description: 'Execute a custom SOQL query to retrieve data from Salesforce',
+export const salesforceQueryMoreTool: ToolConfig<
+  SalesforceQueryMoreParams,
+  SalesforceQueryMoreResponse
+> = {
+  id: 'salesforce_query_more',
+  name: 'Get More Query Results from Salesforce',
+  description: 'Retrieve additional query results using the nextRecordsUrl from a previous query',
   version: '1.0.0',
 
   oauth: {
@@ -47,24 +49,27 @@ export const salesforceQueryTool: ToolConfig<SalesforceQueryParams, SalesforceQu
     accessToken: { type: 'string', required: true, visibility: 'hidden' },
     idToken: { type: 'string', required: false, visibility: 'hidden' },
     instanceUrl: { type: 'string', required: false, visibility: 'hidden' },
-    query: {
+    nextRecordsUrl: {
       type: 'string',
       required: true,
       visibility: 'user-only',
-      description: 'SOQL query to execute (e.g., SELECT Id, Name FROM Account LIMIT 10)',
+      description: 'The nextRecordsUrl from a previous query response',
     },
   },
 
   request: {
     url: (params) => {
-      if (!params.query || params.query.trim() === '') {
+      if (!params.nextRecordsUrl || params.nextRecordsUrl.trim() === '') {
         throw new Error(
-          'SOQL Query is required. Please provide a valid SOQL query (e.g., SELECT Id, Name FROM Account LIMIT 10).'
+          'Next Records URL is required. This should be the nextRecordsUrl value from a previous query response.'
         )
       }
       const instanceUrl = getInstanceUrl(params.idToken, params.instanceUrl)
-      const encodedQuery = encodeURIComponent(params.query)
-      return `${instanceUrl}/services/data/v59.0/query?q=${encodedQuery}`
+      // nextRecordsUrl is typically a relative path like /services/data/v59.0/query/01g...
+      const nextUrl = params.nextRecordsUrl.startsWith('/')
+        ? params.nextRecordsUrl
+        : `/${params.nextRecordsUrl}`
+      return `${instanceUrl}${nextUrl}`
     },
     method: 'GET',
     headers: (params) => ({
@@ -73,15 +78,15 @@ export const salesforceQueryTool: ToolConfig<SalesforceQueryParams, SalesforceQu
     }),
   },
 
-  transformResponse: async (response, params?) => {
+  transformResponse: async (response) => {
     const data = await response.json()
     if (!response.ok) {
       const errorMessage = extractErrorMessage(
         data,
         response.status,
-        'Failed to execute SOQL query'
+        'Failed to get more query results'
       )
-      logger.error('Failed to execute SOQL query', { data, status: response.status })
+      logger.error('Failed to get more query results', { data, status: response.status })
       throw new Error(errorMessage)
     }
 
@@ -94,9 +99,8 @@ export const salesforceQueryTool: ToolConfig<SalesforceQueryParams, SalesforceQu
         totalSize: data.totalSize || records.length,
         done: data.done !== false,
         nextRecordsUrl: data.nextRecordsUrl,
-        query: params?.query || '',
         metadata: {
-          operation: 'query',
+          operation: 'query_more',
           totalReturned: records.length,
           hasMore: !data.done,
         },
@@ -115,7 +119,6 @@ export const salesforceQueryTool: ToolConfig<SalesforceQueryParams, SalesforceQu
         totalSize: { type: 'number', description: 'Total number of records matching query' },
         done: { type: 'boolean', description: 'Whether all records have been returned' },
         nextRecordsUrl: { type: 'string', description: 'URL to fetch next batch of records' },
-        query: { type: 'string', description: 'The executed SOQL query' },
         metadata: { type: 'object', description: 'Operation metadata' },
         success: { type: 'boolean', description: 'Operation success status' },
       },
